fix: remove unnecessary type conversions flagged by clippy
Fix 8 clippy warnings about useless conversions in writer.rs and elf_module.rs:

1. **x86_64/writer.rs** (6 fixes)
   - Remove .try_into().unwrap() from GumCallingConvention (already u32)
   - Remove .try_into().unwrap() from GumArgType constants (already u32)
   - Keep `as u32` casts where constants are i32 but fields expect u32

2. **elf_module.rs** (1 fix)
   - Remove unnecessary cast: (*details).tag already u32

All clippy checks now pass with -D warnings. Verified locally before push.

+++ frida-gum/src/elf_module.rs
@@ -41,7 +41,7 @@ impl ElfDynamicEntry {
     fn from_raw(details: *const GumElfDynamicEntryDetails) -> Self {
         unsafe {
             Self {
-                tag: (*details).tag as u32,
+                tag: (*details).tag,
                 val: (*details).val,
             }
         }
